Move device state buffer initialisation code outside the Device node class.
